requirements.txt
tensorflow>=1.0.0,<2.0.0
tflearn>=0.3.2
pympler>=0.5
cvxopt>=1.1.9
seaborn>=0.8.1
pandas>=0.20.3
ratelimit
pytest-recording